# triton-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Triton Backend",
    "description": "Backend for Triton",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "operationId": "list_users_api_auth_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserDTO"
                  },
                  "type": "array",
                  "title": "Response List Users Api Auth Users Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/users/{user_id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete User",
        "operationId": "delete_user_api_auth_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/users/{user_id}/instances": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update User Instances",
        "operationId": "update_user_instances_api_auth_users__user_id__instances_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserInstancesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/users/{user_id}/role": {
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update User Role",
        "operationId": "update_user_role_api_auth_users__user_id__role_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instances": {
      "post": {
        "tags": [
          "instances"
        ],
        "summary": "Create Instance",
        "description": "Create a Triton instance from UI input (url + optional name).\n\nThe backend connects to Triton, reads the currently available models,\nstores the instance in the DB, and returns the created DTO.",
        "operationId": "create_instance_api_instances_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTritonInstanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TritonInstanceDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "List Instances",
        "description": "List all registered Triton instances.",
        "operationId": "list_instances_api_instances_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TritonInstanceDTO"
                  },
                  "title": "Response List Instances Api Instances Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instances/{instance_id}/s3": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance S3",
        "operationId": "get_instance_s3_api_instances__instance_id__s3_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceS3ConfigDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "instances"
        ],
        "summary": "Update Instance S3",
        "operationId": "update_instance_s3_api_instances__instance_id__s3_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstanceS3Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceS3ConfigDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/s3/list": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "List Instance S3",
        "operationId": "list_instance_s3_api_instances__instance_id__s3_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Prefix"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/s3/content": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance S3 Content",
        "operationId": "get_instance_s3_content_api_instances__instance_id__s3_content_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "S3 object path",
              "title": "Path"
            },
            "description": "S3 object path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3FileContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "instances"
        ],
        "summary": "Put Instance S3 Content",
        "operationId": "put_instance_s3_content_api_instances__instance_id__s3_content_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "S3 object path",
              "title": "Path"
            },
            "description": "S3 object path"
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content Type"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "contentMediaType": "application/octet-stream",
                "title": "Content"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/S3FileWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instances/{instance_id}/s3/content/raw": {
      "get": {
        "tags": [
          "instances"
        ],
        "summary": "Get Instance S3 Content Raw",
        "operationId": "get_instance_s3_content_raw_api_instances__instance_id__s3_content_raw_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Instance Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "S3 object path",
              "title": "Path"
            },
            "description": "S3 object path"
          }
        ],
        "responses": {
          "200": {
            "description": "Raw object bytes",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTritonInstanceRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "verify_ssl": {
            "type": "boolean",
            "title": "Verify Ssl",
            "default": false
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "CreateTritonInstanceRequest",
        "description": "Request to create a Triton instance."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstanceS3ConfigDTO": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bucket"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prefix"
          },
          "access_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Access Key"
          },
          "secret_configured": {
            "type": "boolean",
            "title": "Secret Configured",
            "default": false
          },
          "use_https": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Use Https"
          },
          "verify_ssl": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verify Ssl"
          },
          "ca_certificate": {
            "type": "string",
            "title": "Ca Certificate",
            "default": ""
          },
          "address_style": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "path",
                  "virtual"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Style"
          }
        },
        "type": "object",
        "title": "InstanceS3ConfigDTO"
      },
...
      "S3ListResponse": {
        "properties": {
          "prefix": {
            "type": "string",
            "title": "Prefix"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/S3EntryDTO"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "prefix",
          "entries"
        ],
        "title": "S3ListResponse"
      },
...
      "TritonInstanceDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "model_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Model Names"
          },
          "repository_models": {
            "items": {
              "$ref": "#/components/schemas/TritonRepositoryModelDTO"
            },
            "type": "array",
            "title": "Repository Models"
          },
          "server_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Metadata"
          },
          "health_live": {
            "type": "boolean",
            "title": "Health Live",
            "default": false
          },
          "health_ready": {
            "type": "boolean",
            "title": "Health Ready",
            "default": false
          },
          "health_last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Last Checked At"
          },
          "health_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Error"
          },
          "triton_verify_ssl": {
            "type": "boolean",
            "title": "Triton Verify Ssl",
            "default": false
          },
          "triton_ca_certificate": {
            "type": "string",
            "title": "Triton Ca Certificate",
            "default": ""
          },
          "metrics_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Url"
          },
          "metrics_cpu": {
            "type": "number",
            "title": "Metrics Cpu",
            "default": 0
          },
          "metrics_ram": {
            "type": "number",
            "title": "Metrics Ram",
            "default": 0
          },
          "metrics_gpu": {
            "type": "number",
            "title": "Metrics Gpu",
            "default": 0
          },
          "metrics_last_checked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Last Checked At"
          },
          "metrics_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics Error"
          },
          "deployment_runtime": {
            "type": "string",
            "title": "Deployment Runtime",
            "default": "external"
          },
          "deployment_namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Namespace"
          },
          "deployment_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Name"
          },
          "deployment_service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Service Name"
          },
          "deployment_secret_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployment Secret Name"
          },
          "deployment_log": {
            "type": "string",
            "title": "Deployment Log",
            "default": ""
          },
          "is_self_deployed": {
            "type": "boolean",
            "title": "Is Self Deployed",
            "default": false
          },
          "pod_statuses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pod Statuses"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "s3": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstanceS3ConfigDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "name",
          "model_names",
          "created_at"
        ],
        "title": "TritonInstanceDTO",
        "description": "DTO for outgoing Triton instance data (API responses)."
      },
...
      "UpdateUserInstancesRequest": {
        "properties": {
          "assigned_instances": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assigned Instances"
          }
        },
        "type": "object",
        "title": "UpdateUserInstancesRequest"
      },
      "UpdateUserRoleRequest": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "UpdateUserRoleRequest"
      },
      "UserDTO": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "auth_provider": {
            "type": "string",
            "title": "Auth Provider"
          },
          "oidc_subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oidc Subject"
          },
          "assigned_instances": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assigned Instances"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "role",
          "auth_provider",
          "assigned_instances",
          "is_active",
          "created_at"
        ],
        "title": "UserDTO"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}